requirements.txt
aiogram
pydantic
python-docx
asyncio
requests
pydantic_settings
aiohttp
comtypes
redis
beanie
aioredis